+++ CHANGES.md
@@ -230,4 +230,5 @@
 - Starkey: fix timezone bug
 - Quick Scan: add cool down warning
 - Quick Scan: fix timezone bug
+- Extra Menu: show last timestamp performing Quick Scan
 - Underhood bug fix